ソースディレクトリ構成を見直しリファクタリング
