+++ components/manage/content/contact.vue
@@ -100,7 +100,7 @@ const onCancel = () => {
           placeholder="本文を入力してください"
         />
       </div>
-      <EditorFormActions
+      <ManageContentFormActions
         :contentId="contactData?.id"
         class="mt-4 mb-2"
         @create="onCreate"
