+++ components/manage/content/image-setting.vue
