+++ components/manage/content/news-whole-switch.vue
