+++ components/manage/content/service.vue
@@ -101,7 +101,7 @@ const onCancel = () => {
           placeholder="本文を入力してください"
         />
       </div>
-      <EditorFormActions
+      <ManageContentFormActions
         :contentId="serviceData?.id"
         class="mt-4 mb-2"
         @create="onCreate"
