+++ components/manage/customer/home-layout-setting.vue
