+++ components/manage/customer/on-preview.vue
