+++ components/publish/home/type1/information.vue
@@ -1,93 +1,76 @@
 <script setup lang="ts">
 const { customerId, canEdit } = useFoundation()
 const {
-  newsRef,
+  informationRef,
   onLoad,
   onCreate,
   onUpdate,
   onRemove,
   onUpdateImageSetting,
   loading,
-} = useNewsActions(customerId)
-
-const route = useRoute()
-const contentId = route.params.id
-await onLoad(contentId)
+} = useInformationActions(customerId)
+await onLoad()
 </script>
 
 <template>
-  <CommonContentCard class="news-detail">
-    <CommonContentWrap :loading="loading">
+  <CommonContentWrap :loading="loading">
+    <CommonContentCard class="type1-information">
       <CommonEyecatchImage
-        v-if="newsRef?.image"
-        :url="newsRef?.image?.url"
-        :settings="newsRef?.image?.settings"
+        v-if="informationRef?.image"
+        :url="informationRef?.image?.url"
+        :settings="informationRef?.image?.settings"
         class="eyecatcher"
       >
-        <div v-if="canEdit && newsRef?.image?.settings" class="image-settings">
-          <EditorImageSetting
-            :settings="newsRef.image.settings"
+        <CommonEyecatchTitles
+          place="section"
+          :title="informationRef?.title"
+          :subtitle="informationRef?.subtitle"
+          class="eyecatcher__titles"
+        />
+        <div
+          v-if="canEdit && informationRef?.image?.settings"
+          class="image-settings"
+        >
+          <ManageContentImageSetting
+            :settings="informationRef.image.settings"
             @update="onUpdateImageSetting"
           />
         </div>
       </CommonEyecatchImage>
       <CommonContentCardBody>
-        <template v-if="newsRef?.id">
-          <div class="news-detail__header">
-            <CommonNewsCategoryBadge :category="newsRef.category" small />
-            <p>
-              <small>{{
-                formatLocalDate(newsRef.publishOn, 'YYYY/MM/DD')
-              }}</small>
-            </p>
+        <div v-if="informationRef?.body">
+          <div v-html="htmlSanitizer(informationRef?.body)" />
+          <div class="inquire-activator">
+            <PublishInquire />
           </div>
-          <h5 class="g-text-cl news-detail__title">
-            <span>{{ newsRef?.title ?? '' }}</span>
-          </h5>
-          <div v-if="newsRef?.body">
-            <div v-html="htmlSanitizer(newsRef?.body)" />
-            <div class="inquire-activator">
-              <ScreenInquire />
-            </div>
-          </div>
-        </template>
+        </div>
         <div v-else class="no-items">
           <p>データがありません</p>
           <div v-if="canEdit">
-            <EditorNews
+            <ManageContentInformation
               activaterLabel="コンテンツを登録してください"
               @create="onCreate"
             />
           </div>
         </div>
       </CommonContentCardBody>
-      <div v-if="canEdit && newsRef?.id" class="edit-activator">
-        <EditorNews
-          :news-data="newsRef"
+      <div v-if="canEdit && informationRef?.id" class="edit-activator">
+        <ManageContentInformation
+          :information-data="informationRef"
           @update="onUpdate"
           @remove="onRemove"
         />
       </div>
-    </CommonContentWrap>
-  </CommonContentCard>
+    </CommonContentCard>
+  </CommonContentWrap>
 </template>
 
 <style scoped lang="scss">
 $eyecatcher-height: 500px;
 $eyecatcher-height-sm: 600px;
 
-.news-detail {
+.type1-information {
   position: relative;
-  &__header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  &__title {
-    font-size: 1.5rem;
-    font-weight: bold;
-    margin-bottom: 1.8rem;
-  }
   .edit-activator {
     position: absolute;
     top: 1rem;
