+++ components/publish/home/type1/section-title.vue
