+++ components/publish/home/type1/services.vue
@@ -46,7 +46,7 @@ await onLoad()
               v-html="htmlSanitizer((content as ServiceType).caption)"
             />
             <div class="edit-activator">
-              <EditorService
+              <ManageContentService
                 :serviceData="content as ServiceType"
                 activaterSize="x-small"
                 @update="onUpdate"
@@ -59,14 +59,14 @@ await onLoad()
       <div v-else class="no-items">
         <p>データがありません</p>
         <div>
-          <EditorService
+          <ManageContentService
             activaterLabel="コンテンツを登録してください"
             @create="onCreate"
           />
         </div>
       </div>
       <div v-if="serviceListRef?.length" class="create-activator">
-        <EditorService @create="onCreate" />
+        <ManageContentService @create="onCreate" />
       </div>
     </template>
     <template v-else>
