+++ components/publish/inquire/result.vue
