+++ components/publish/news/detail.vue
@@ -1,76 +1,93 @@
 <script setup lang="ts">
 const { customerId, canEdit } = useFoundation()
 const {
-  informationRef,
+  newsRef,
   onLoad,
   onCreate,
   onUpdate,
   onRemove,
   onUpdateImageSetting,
   loading,
-} = useInformationActions(customerId)
-await onLoad()
+} = useNewsActions(customerId)
+
+const route = useRoute()
+const contentId = route.params.id
+await onLoad(contentId)
 </script>
 
 <template>
-  <CommonContentWrap :loading="loading">
-    <CommonContentCard class="type1-information">
+  <CommonContentCard class="news-detail">
+    <CommonContentWrap :loading="loading">
       <CommonEyecatchImage
-        v-if="informationRef?.image"
-        :url="informationRef?.image?.url"
-        :settings="informationRef?.image?.settings"
+        v-if="newsRef?.image"
+        :url="newsRef?.image?.url"
+        :settings="newsRef?.image?.settings"
         class="eyecatcher"
       >
-        <CommonEyecatchTitles
-          place="section"
-          :title="informationRef?.title"
-          :subtitle="informationRef?.subtitle"
-          class="eyecatcher__titles"
-        />
-        <div
-          v-if="canEdit && informationRef?.image?.settings"
-          class="image-settings"
-        >
-          <EditorImageSetting
-            :settings="informationRef.image.settings"
+        <div v-if="canEdit && newsRef?.image?.settings" class="image-settings">
+          <ManageContentImageSetting
+            :settings="newsRef.image.settings"
             @update="onUpdateImageSetting"
           />
         </div>
       </CommonEyecatchImage>
       <CommonContentCardBody>
-        <div v-if="informationRef?.body">
-          <div v-html="htmlSanitizer(informationRef?.body)" />
-          <div class="inquire-activator">
-            <ScreenInquire />
+        <template v-if="newsRef?.id">
+          <div class="news-detail__header">
+            <CommonNewsCategoryBadge :category="newsRef.category" small />
+            <p>
+              <small>{{
+                formatLocalDate(newsRef.publishOn, 'YYYY/MM/DD')
+              }}</small>
+            </p>
           </div>
-        </div>
+          <h5 class="g-text-cl news-detail__title">
+            <span>{{ newsRef?.title ?? '' }}</span>
+          </h5>
+          <div v-if="newsRef?.body">
+            <div v-html="htmlSanitizer(newsRef?.body)" />
+            <div class="inquire-activator">
+              <PublishInquire />
+            </div>
+          </div>
+        </template>
         <div v-else class="no-items">
           <p>データがありません</p>
           <div v-if="canEdit">
-            <EditorInformation
+            <ManageContentNews
               activaterLabel="コンテンツを登録してください"
               @create="onCreate"
             />
           </div>
         </div>
       </CommonContentCardBody>
-      <div v-if="canEdit && informationRef?.id" class="edit-activator">
-        <EditorInformation
-          :information-data="informationRef"
+      <div v-if="canEdit && newsRef?.id" class="edit-activator">
+        <ManageContentNews
+          :news-data="newsRef"
           @update="onUpdate"
           @remove="onRemove"
         />
       </div>
-    </CommonContentCard>
-  </CommonContentWrap>
+    </CommonContentWrap>
+  </CommonContentCard>
 </template>
 
 <style scoped lang="scss">
 $eyecatcher-height: 500px;
 $eyecatcher-height-sm: 600px;
 
-.type1-information {
+.news-detail {
   position: relative;
+  &__header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+  &__title {
+    font-size: 1.5rem;
+    font-weight: bold;
+    margin-bottom: 1.8rem;
+  }
   .edit-activator {
     position: absolute;
     top: 1rem;
